+++ R/data.R
@@ -1,11 +1,11 @@
 #' @name words
 #' @title Common English Words
-#' @description A subset of 283 of the 1000 most commong English words.
+#' @description A subset of 283 of the 1000 most common English words.
 #' From the original list, only words longer than 5 and without apostrophes were selected.
 #' @docType data
 #' @keywords dataset
 #' @usage words
-#' @note I should have just used \code{stringr::\link[stringr]{words}}
+#' @note I should have just used `stringr::words`
 #' @source <https://gist.github.com/deekayen/4148741/01c6252ccc5b5fb307c1bb899c95989a8a284616/>
 #' @format A `character` of length 283 containing words.
 "words"