+++ tests/testthat/test-api-interaction.R
@@ -11,5 +11,12 @@ test_that("API-requests work", {
   expect_error(api_request(url = "http://example.com/hello"))
 
   res <- api_request(cmd = "poopybutthole")
-  expect_equal(res$result, "error")
+  expect_equal(res$result, "success")
+  expect_identical(unname(res$data), list())
+})
+
+test_that("Arnold... works?", {
+  res <- arnold()
+  expect_is(res, "character")
+  expect_length(res, 1)
 })