+++ NEWS.md
@@ -1,8 +1,24 @@
+# tauturri 0.3.0
+
+- Add `search_server` (`search`-API Method)
+- Add `update_check` and `update_tautulli` for **Tautulli** updates
+- Add `get_pms_update` for **Plex Media Server** updates
+- Minor documentation fixes
+- Fix issues due to changes in the API
+
+# tauturri 0.2.0
+
+- Add `get_user_ips`.
+- Add `get_user_logins`.
+- Add `get_server_friendly_name`.
+- Add `get_server_id`.
+- Add `arnold()`, probably the most crucial function of them all.
+
 # tauturri 0.1.1
 
 - Add `get_home_stats`. It was nasty.
-- Apply `use_tidy_style()`
-- Add GitHub links to `DESCRIPTION`
+- Apply `use_tidy_style()`.
+- Add GitHub links to `DESCRIPTION`.
 
 # tauturri 0.1.0